move : PartyIdRequest common 으로 이동
