+++ src/main/java/com/playus/twpservice/domain/common/PartyIdRequest.java
@@ -1,4 +1,4 @@
-package com.playus.twpservice.domain.party.dto.partyupdate;
+package com.playus.twpservice.domain.common;
 
 import jakarta.validation.constraints.Min;
